--- component/http/v2/router/httprouter/metric.go
@@ -1,14 +1,0 @@
-package httprouter
-
-import (
-	"github.com/prometheus/client_golang/prometheus/promhttp"
-)
-
-const (
-	metricsPath = "/metrics"
-)
-
-func metricRoute() *Route {
-	route, _ := NewRecoveryGetRoute(metricsPath, promhttp.Handler().ServeHTTP)
-	return route
-}
